1.2.11

+++ handlers/message.js
@@ -22,7 +22,7 @@ module.exports = async (ctx) => {
     const lngDetector = new LanguageDetect()
     const detect = lngDetector.detect(ctx.message.text)
 
-    if (detect.length > 0) {
+    if (detect.length > 0 && detect[0][1] > 0.3) {
       if (ctx.groupInfo.settings.removeLng.indexOf(detect[0][0]) >= 0) {
         ctx.deleteMessage()
       }
